A service that can be used for dependancy injection to switch instances
We can bound this service to a container in web.php and switch whether we are calling the Book.php model or BookAdapter.php. It is called a dependancy injection and it is very useful because you have to write a single line of code in order to switch the instances.